refactor(tunnel-store): group flat root files into db/ and config/ subdirs
- sqlite.rs + sqlite_rules.rs → db/sqlite.rs + db/rules.rs (db adapters belong together)
- server_config.rs → config/mod.rs (config parsing is an edge adapter)
- lib.rs keeps sqlite, sqlite_rules, server_config as compat aliases
- open_sqlite_pool re-exported from db::sqlite at crate root

cargo check -p tunnel-store: ok, cargo check --workspace: ok
