+++ tunnel-store/src/db/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(feature = "sqlite")]
+pub mod sqlite;
+#[cfg(feature = "sqlite")]
+pub mod rules;
